api/requirements.txt
langgraph
langchain
langchain-openai
langchain-community
langchain-core
langchain-text-splitters
langchain-google-genai
langchain-ollama
python-multipart
fastapi
uvicorn[standard]
sqlmodel
sqlalchemy-stubs
httpx
pytest
pytest-asyncio
pytest-cov
pytest-mock
mypy
psycopg2-binary
asyncpg
alembic
pydub
numpy
openai-whisper
websockets
duckduckgo-search
langchain-community